.claude/skills/agent-scaffolding-toolkit/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-scaffolding-toolkit"
version = "1.0.0"
description = "Surgical toolkit for creating specialized agents with interactive wizard-guided generation"
authors = [{name = "Agent Scaffolding Team"}]
license = {text = "MIT"}
readme = "SKILL.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pyyaml>=6.0",
    "argparse",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
]

[project.scripts]
agent-scaffolder = "create_agent:main"
agent-validator = "validate_agent:main"
agent-lister = "list_templates:main"
agent-enhancer = "enhance_agent:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"*" = ["*.py", "*.json", "*.md"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false